ai-service/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
langchain
langgraph
langchain-google-genai
langchain-postgres
langchain_community
duckduckgo-search
newspaper3k
requests
ddgs
google.generativeai
pypdf
openpyxl
langchain-postgres
pgvector
psycopg2-binary
psycopg2
